zomato normalizer refactored
