+++ src/lib/zomato/stop-words.json
@@ -1,0 +1,18 @@
+[
+  "^POLÉVKA$",
+  "HLAVNÍ JÍDLA",
+  "DNES PO DOBU",
+  "S MOŽNOSTÍ ZAKOUPENÍ",
+  "SALÁTY a MALÁ LEHKÁ JÍDLA",
+  "Informaci o alergenech",
+  "Dle denní nabídky",
+  "^menu$",
+  "^SALÁT$",
+  "^SPECIALITA$",
+  "^Specialita dne$",
+  "^DEZERT$",
+  "^HLAVNÍ CHOD$",
+  "^Bezmasé jídlo$",
+  "^Stálá nabídka$",
+  "^NÁPOJ K POLEDNÍMU MENU$"
+]
